[feat] 404 Error 응답 추가
- 작성한 API 디자인에 404 에러 응답 추가
- 에러 글로벌 핸들링으로 변경
